CI: fix rust formatting
`cargo fmt --all` should be avoided in CI as that will include `rust/ast-generator`
which has sources provided by bazel (`bazel run //rust/ast-generator:inject_sources`
can provide those sources in-tree).

Now the formatting checks are limited to the sources that trigger the jobs, and a
check is added to `rust/ast-generator`.