Code scanning config: Exclude actions test directory
These are test cases for the GitHub Actions analysis.
Exclude them when running code scanning against this repo,
to avoid noisy alerts.

Test workflow files in this directory are safe from
execution, because Actions only executes workflows
that live directly in the .github/workflows
top-level directory.

`action.yml` files in this directory can in theory
be executed as a step in a workflow; for now exclude them.