Misc: Add script creating DCA source suites from MRVA
The script takes the URL of a MRVA exported Gist and uses it to produce
a source suite compatible with DCA.

At present, you have to manually specify the language on the
commandline, using the `--language` parameter.

Also supports `--min` and `--max` parameters if you want to limit the
sources to ones with a bounded number of alerts.