Rust: skip unexpanded stuff in library emission
This will skip all unexpanded entities in library extraction, where we
only really care about expanded things. This means skipping:

* the token tree of macro calls
* the unexpanded AST of attribute macros

In the latter case, in order to replace the single `Item` with its
expansion (which is a `MacroItems` entity), we wrap the `MacroItems` in
a dummy `MacroCall` with null path.